Figured out that the issue is in the GDT

+++ kernel/src/Memory/Memory.c
@@ -6,7 +6,7 @@ uint64_t GetMemorySize(EFI_MEMORY_DESCRIPTOR* memoryMap, uint64_t memoryMapEntri
     if (memorySizeBytes > 0)
         return memorySizeBytes;
 
-    for (int i = 0; i < memoryMapEntries; ++i)
+    for (uint64_t i = 0; i < memoryMapEntries; ++i)
     {
         EFI_MEMORY_DESCRIPTOR* descriptor = (EFI_MEMORY_DESCRIPTOR*)((uint64_t)memoryMap+(i*descriptorSize));
         memorySizeBytes += descriptor->pagesAmount * 4096;
